// src/services/universalFile/types/api.ts
export interface FolderCreateParams {
  /** 文件夹名称 */
  name: string;
  /** 父文件夹ID */
  parentId?: string;
  /** 模块ID */
  moduleId?: string;
  /** 业务ID */
  businessId?: string;
  /** 文件夹描述 */
  description?: string;
  /** 显示顺序 */
  sortOrder?: number;
}
// ...
export interface FolderUpdateParams {
  /** 文件夹名称 */
  name?: string;
  /** 父文件夹ID */
  parentId?: string;
  /** 文件夹描述 */
  description?: string;
  /** 显示顺序 */
  sortOrder?: number;
}